backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
supabase==2.27.2
python-dotenv==1.2.1
pydantic==2.12.5
openai==1.59.4
sentence-transformers==5.2.2
faiss-cpu==1.13.2
pypdf==6.6.2
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==12.1.0
